[style] #92 리뷰 반영 - UiModel 분리
