+++ presentation/src/main/java/com/mashup/presentation/detail/chat/compose/Chat.kt
@@ -20,6 +20,8 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.mashup.presentation.detail.chat.model.MessageBackgroundColor
+import com.mashup.presentation.detail.chat.model.MessageUiModel
 import com.mashup.presentation.ui.theme.*
 
 /**
